chore: moving type files to sub-modules
