+++ src/leapfrogai_api/backend/grpc_client.py
@@ -3,26 +3,31 @@
 from typing import Iterator, AsyncGenerator, Any, List
 import grpc
 from fastapi.responses import StreamingResponse
+
 import leapfrogai_sdk as lfai
 from leapfrogai_api.backend.helpers import recv_chat, recv_completion
-from leapfrogai_api.backend.types import (
-    ChatChoice,
-    ChatCompletionResponse,
-    ChatMessage,
-    CompletionChoice,
-    CompletionResponse,
-    CreateEmbeddingResponse,
-    EmbeddingResponseData,
-    Usage,
-)
-from leapfrogai_api.typedef.requests import (
-    CreateTranscriptionResponse,
-    CreateTranslationResponse,
-)
 from leapfrogai_sdk.chat.chat_pb2 import (
     ChatCompletionResponse as ProtobufChatCompletionResponse,
 )
 from leapfrogai_api.utils.config import Model
+from leapfrogai_api.typedef.audio import (
+    CreateTranscriptionResponse,
+    CreateTranslationResponse,
+)
+from leapfrogai_api.typedef.chat import (
+    ChatChoice,
+    ChatCompletionResponse,
+    ChatMessage,
+)
+from leapfrogai_api.typedef.completion import (
+    CompletionChoice,
+    CompletionResponse,
+)
+from leapfrogai_api.typedef import Usage
+from leapfrogai_api.typedef.embeddings import (
+    CreateEmbeddingResponse,
+    EmbeddingResponseData,
+)
 
 
 async def stream_completion(model: Model, request: lfai.CompletionRequest):
