+++ src/leapfrogai_api/routers/openai/assistants.py
@@ -2,8 +2,9 @@
 
 from fastapi import HTTPException, APIRouter, status
 from openai.types.beta import Assistant, AssistantDeleted
+
 from leapfrogai_api.backend.helpers import object_or_default
-from leapfrogai_api.typedef import ListAssistantsResponse
+from leapfrogai_api.typedef.assistants import ListAssistantsResponse
 from leapfrogai_api.routers.openai.requests.create_modify_assistant_request import (
     CreateAssistantRequest,
     ModifyAssistantRequest,
