+++ src/leapfrogai_api/routers/openai/audio.py
@@ -3,9 +3,10 @@
 from itertools import chain
 from typing import Annotated
 from fastapi import HTTPException, APIRouter, Depends
+
 from leapfrogai_api.backend.grpc_client import create_transcription, create_translation
 from leapfrogai_api.backend.helpers import read_chunks
-from leapfrogai_api.typedef.requests import (
+from leapfrogai_api.typedef.audio import (
     CreateTranscriptionRequest,
     CreateTranscriptionResponse,
     CreateTranslationRequest,
