+++ src/leapfrogai_api/routers/openai/completions.py
@@ -6,9 +6,7 @@ from leapfrogai_api.backend.grpc_client import (
     completion,
     stream_completion,
 )
-from leapfrogai_api.backend.types import (
-    CompletionRequest,
-)
+from leapfrogai_api.typedef.completion import CompletionRequest
 from leapfrogai_api.routers.supabase_session import Session
 from leapfrogai_api.utils import get_model_config
 from leapfrogai_api.utils.config import Config
