+++ src/leapfrogai_api/routers/openai/embeddings.py
@@ -1,11 +1,14 @@
 """FastAPI router for OpenAI embeddings API."""
 
 from typing import Annotated
-
 from fastapi import APIRouter, Depends, HTTPException, status
+
 import leapfrogai_sdk as lfai
 from leapfrogai_api.backend.grpc_client import create_embeddings
-from leapfrogai_api.backend.types import CreateEmbeddingRequest, CreateEmbeddingResponse
+from leapfrogai_api.typedef.embeddings import (
+    CreateEmbeddingRequest,
+    CreateEmbeddingResponse,
+)
 from leapfrogai_api.routers.supabase_session import Session
 from leapfrogai_api.utils import get_model_config
 from leapfrogai_api.utils.config import Config
