+++ src/leapfrogai_api/routers/openai/runs.py
@@ -5,9 +5,7 @@ from fastapi import HTTPException, APIRouter, status
 from fastapi.responses import StreamingResponse
 from openai.types.beta.threads import Run
 from openai.pagination import SyncCursorPage
-from leapfrogai_api.routers.openai.requests.thread_run_create_params_request import (
-    ThreadRunCreateParamsRequest,
-)
+
 from leapfrogai_api.data.crud_run import CRUDRun
 from leapfrogai_api.data.crud_thread import CRUDThread
 from leapfrogai_api.routers.supabase_session import Session
@@ -15,7 +13,8 @@ from leapfrogai_api.utils.validate_tools import (
     validate_assistant_tool,
     validate_assistant_tool_choice_option,
 )
-from leapfrogai_api.typedef.requests import RunCreateParamsRequest, ModifyRunRequest
+from leapfrogai_api.typedef.threads import ThreadRunCreateParamsRequest
+from leapfrogai_api.typedef.runs import RunCreateParamsRequest, ModifyRunRequest
 from leapfrogai_api.backend.thread_runner import ThreadRunner
 
 router = APIRouter(prefix="/openai/v1/threads", tags=["openai/threads/runs"])
