+++ src/leapfrogai_api/typedef/assistants/__init__.py
@@ -1,0 +1,1 @@
+from .assistant_types import ListAssistantsResponse as ListAssistantsResponse
