+++ src/leapfrogai_api/typedef/audio/__init__.py
@@ -1,0 +1,6 @@
+from .audio_types import (
+    CreateTranscriptionRequest as CreateTranscriptionRequest,
+    CreateTranscriptionResponse as CreateTranscriptionResponse,
+    CreateTranslationRequest as CreateTranslationRequest,
+    CreateTranslationResponse as CreateTranslationResponse,
+)
