+++ src/leapfrogai_api/typedef/auth/__init__.py
@@ -1,0 +1,4 @@
+from .auth_types import (
+    CreateAPIKeyRequest as CreateAPIKeyRequest,
+    ModifyAPIKeyRequest as ModifyAPIKeyRequest,
+)
