+++ src/leapfrogai_api/typedef/completion/__init__.py
@@ -1,0 +1,5 @@
+from .completion_types import (
+    CompletionChoice as CompletionChoice,
+    CompletionRequest as CompletionRequest,
+    CompletionResponse as CompletionResponse,
+)
