+++ src/leapfrogai_api/typedef/embeddings/__init__.py
@@ -1,0 +1,5 @@
+from .embedding_types import (
+    EmbeddingResponseData as EmbeddingResponseData,
+    CreateEmbeddingRequest as CreateEmbeddingRequest,
+    CreateEmbeddingResponse as CreateEmbeddingResponse,
+)
